+++ tests/schema/types/test_uuid.py
@@ -3,6 +3,7 @@ import uuid
 from graphql import GraphQLError
 
 import strawberry
+from strawberry.utils import IS_GQL_32
 
 
 def test_uuid():
@@ -63,7 +64,11 @@ def test_serialization_of_incorrect_uuid_string():
 
     assert result.errors
     assert isinstance(result.errors[0], GraphQLError)
-    assert result.errors[0].message == (
+    expected_message = (
         "Variable '$value' got invalid value 'fail'; Value cannot represent a "
         'UUID: "fail". badly formed hexadecimal UUID string'
+        if IS_GQL_32
+        else "Variable '$value' has invalid value: Value cannot represent a "
+        'UUID: "fail". badly formed hexadecimal UUID string'
     )
+    assert result.errors[0].message == expected_message